LayoutTests/fast/canvas/canvas-strokeText-strokes-shadow-expected.txt

Test that strokeText() doesn't produce a filled shadow.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS imgdata[0] is 0
PASS imgdata[1] is 128
PASS imgdata[2] is 0
PASS imgdata[3] is 255
PASS successfullyParsed is true

TEST COMPLETE
